+++ packages/beacon-node/src/chain/archiveStore/archiveBlocks.ts
