+++ packages/beacon-node/src/chain/archiveStore/archiver.ts
