+++ packages/beacon-node/src/chain/archiveStore/index.ts
