+++ packages/beacon-node/src/chain/archiveStore/interface.ts
