+++ packages/beacon-node/src/chain/archiveStore/strategies/frequencyStateArchiveStrategy.ts
