+++ packages/beacon-node/src/chain/options.ts
@@ -1,15 +1,15 @@
 import {SAFE_SLOTS_TO_IMPORT_OPTIMISTICALLY} from "@lodestar/params";
 import {defaultOptions as defaultValidatorOptions} from "@lodestar/validator";
-import {DEFAULT_STATE_ARCHIVE_MODE} from "./archiver/archiver.js";
-import {ArchiverOpts} from "./archiver/interface.js";
+import {DEFAULT_STATE_ARCHIVE_MODE} from "./archiveStore/archiver.js";
+import {ArchiverOpts} from "./archiveStore/interface.js";
 import {ForkChoiceOpts} from "./forkChoice/index.js";
 import {LightClientServerOpts} from "./lightClient/index.js";
 import {ShufflingCacheOpts} from "./shufflingCache.js";
 import {DEFAULT_MAX_BLOCK_STATES, FIFOBlockStateCacheOpts} from "./stateCache/fifoBlockStateCache.js";
 import {PersistentCheckpointStateCacheOpts} from "./stateCache/persistentCheckpointsCache.js";
 import {DEFAULT_MAX_CP_STATE_EPOCHS_IN_MEMORY} from "./stateCache/persistentCheckpointsCache.js";
-export {StateArchiveMode} from "./archiver/interface.js";
-export {DEFAULT_STATE_ARCHIVE_MODE} from "./archiver/archiver.js";
+export {StateArchiveMode} from "./archiveStore/interface.js";
+export {DEFAULT_STATE_ARCHIVE_MODE} from "./archiveStore/archiver.js";
 
 export type IChainOptions = BlockProcessOpts &
   PoolOpts &
