+++ packages/beacon-node/test/unit/chain/archiveStore/stateArchiver.test.ts
@@ -1,6 +1,6 @@
 import {computeStartSlotAtEpoch} from "@lodestar/state-transition";
 import {describe, expect, it} from "vitest";
-import {computeStateSlotsToDelete} from "../../../../src/chain/archiver/strategies/frequencyStateArchiveStrategy.js";
+import {computeStateSlotsToDelete} from "../../../../src/chain/archiveStore/strategies/frequencyStateArchiveStrategy.js";
 
 describe("state archiver task", () => {
   describe("computeStateSlotsToDelete", () => {
